backend/scheduler/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
apscheduler==3.8.1
httpx==0.19.0
pymongo==3.12.0
python-dotenv==0.19.0